refactoring for loop stmt and interfaces
